requirements.txt
notion-client
undetected-chromedriver
bs4
python-telegram-bot
transformers
torchvision
